refactor: Main 컴포넌트 폴더 생성
